[Update] Repository

+++ src/main/java/sit/project/intregratedbackend/repositories/AccountsRepository.java
@@ -2,6 +2,6 @@ package sit.project.intregratedbackend.repositories;
 import sit.project.intregratedbackend.models.Accounts;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface AccountsRepository extends JpaRepository<Accounts, Long>{
+public interface AccountsRepository extends JpaRepository<Accounts, Integer>{
 	
 }
