+++ src/main/java/sit/project/intregratedbackend/repositories/CategoriesRepository.java
@@ -3,6 +3,6 @@ package sit.project.intregratedbackend.repositories;
 import org.springframework.data.jpa.repository.JpaRepository;
 import sit.project.intregratedbackend.models.Categories;
 
-public interface CategoriesRepository extends JpaRepository<Categories, Long>{
+public interface CategoriesRepository extends JpaRepository<Categories, Integer>{
 
 }
